+++ Amazon.Pay.API.SDK/WebStore/CheckoutSession/CreateCheckoutSessionRequest.cs
@@ -30,6 +30,21 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
         }
 
         /// <summary>
+        /// Initializes a new instance of the CreateCheckoutSessionRequest class.
+        /// </summary>
+        /// <param name="checkoutReviewReturnUrl">Checkout review URL provided by the merchant. Amazon Pay will redirect to this URL after the buyer selects their preferred payment instrument and shipping address.</param>
+        /// <param name="storeId">Store ID as defined in Seller Central.</param>
+        /// <param name="scopes">Scopes passed by merchant to request buyer data</param>
+        public CreateCheckoutSessionRequest(string checkoutReviewReturnUrl, string storeId, CheckoutSessionScope[] checkoutSessionScopes) : base()
+        {
+            WebCheckoutDetails.CheckoutReviewReturnUrl = checkoutReviewReturnUrl;
+            StoreId = storeId;
+            DeliverySpecifications = new DeliverySpecifications();
+            AddressDetails = new AddressDetails();
+            CheckoutSessionScope = checkoutSessionScopes;
+        }
+
+        /// <summary>
         /// Login with Amazon client ID. Do not use the application ID.
         /// </summary>
         [JsonProperty(PropertyName = "storeId")]
@@ -46,6 +61,12 @@ namespace Amazon.Pay.API.WebStore.CheckoutSession
         /// </summary>
         [JsonProperty(PropertyName = "addressDetails")]
         public AddressDetails AddressDetails { get; internal set; }
+
+        /// <summary>
+        /// Checkout Session Scopes
+        /// </summary>
+        [JsonProperty(PropertyName = "scopes")]
+        public CheckoutSessionScope[] CheckoutSessionScope { get; internal set; }
 
     }
 }